requirements.txt
certifi==2019.3.9
chardet==3.0.4
chromedriver-binary==75.0.3770.8.0
cycler==0.10.0
decorator==4.4.0
idna==2.8
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==3.1.0
networkx==2.3
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
pygeocoder==1.2.5
pyocr==0.7
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
